week2/requirements.txt
# AILearn — root-level reference / shared dependencies
# Prefer per-project install: see week1/ragproject, week2/project6-research-agent for their requirements.txt
# project1 and project2 use pyproject.toml (pip install -e . or uv sync)

# LangChain
langchain>=0.2.0,<0.3.0
langchain-community>=0.2.0,<0.3.0
langchain-core>=0.2.0,<0.3.0
langchain-ollama>=1.0.0
langgraph>=0.2.0,<0.3.0

# Optional: RAG / week1
langchain-huggingface

# Vector store (week1, project6)
chromadb>=0.5.0
sentence-transformers>=3.0.0

# Data validation
pydantic>=2.8.0,<3.0.0
pydantic-settings>=2.4.0,<3.0.0

# APIs & I/O
arxiv>=2.0.0
httpx>=0.27.0
pypdf>=4.0.0
python-dotenv>=1.0.0
requests>=2.32.0

# LLM
ollama>=0.3.0

# Utilities
rich
